minor changes in problem instance classes ( vehicle speed added, result as double)

+++ ComputationalCluster/AlgorithmSolvers/DVRPEssentials/DVRPPartialProblemInstance.cs
@@ -19,7 +19,7 @@ namespace AlgorithmSolvers.DVRPEssentials
         [Description("Miejsca do odwiedzenia - klienci")]
         public int[][] VisitIds { get; set; }
         [Description("Wynik problemu - suma minimalnych odległości ścieżek odwiedzających wszystkich klientów")]
-        public int PartialResult { get; set; }
+        public double PartialResult { get; set; }
         public SolutionResult SolutionResult { get; set; } = SolutionResult.NotSolved;
 
         public DVRPPartialProblemInstance()
